refactor: トランスパイルエラーになりそうなところを直していく #21
