+++ app/root-post/dto/root-post-dto.ts
@@ -1,12 +1,12 @@
 import { RootPost } from "../../../domain/root-post/entity/root-post";
-import type { Status } from "../../../domain/root-post/entity/root-post";
+import type { Status } from "../../../domain/root-post/value-object/status";
 
 export class RootPostDTO {
   public readonly id: string;
   public readonly content: string;
   public readonly status: Status;
-  public readonly teamId: number;
-  public readonly userId: number;
+  public readonly teamId: string;
+  public readonly userId: string;
 
   public constructor(rootPost: RootPost) {
     this.id = rootPost.post.id;
