+++ domain/root-post/value-object/status.ts
