+++ infra/clerk-user-repository-impl.ts
@@ -16,6 +16,7 @@ export default class ClerkUserRepoImpl {
     return new ClerkUser({
       id: randomId(),
       teamId,
+      isAdmin: false // 仮
     });
   }
 }
